requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==20.1.0